refactor(smn): move smn resources to services path
